+++ src/lib/ui/Modal.svelte
@@ -14,7 +14,7 @@
 
 <!-- Main modal -->
 <div>
-  <div class="bg-black bg-opacity-80 fixed inset-0 z-40" />
+  <div class="bg-black bg-opacity-80 backdrop-blur-[20px] fixed inset-0 z-40" />
   <div
     tabindex="-1"
     class="fixed top-0 left-0 right-0 z-50 w-full overflow-x-hidden overflow-y-auto md:inset-0 h-modal md:h-full justify-center items-center flex"
